requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow==11.2.1
Flask==2.2.3
requests==2.28.1
beautifulsoup4==4.12.2
numpy>=1.26.4
pandas>=2.0.0
Flask‑SQLAlchemy==3.0.3
Flask‑Login==0.6.4
